Move CST -> Types conversion methods to the collect types pass
